Create BaseQuery and have SearchQuery & MLTQuery extend it
